+++ Assets/Dust/Experimental/Scripts/Runtime/Variables/AbstractVariable.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Dust.Experimental.Variables
+{
+    public abstract class AbstractVariable : MonoBehaviour
+    {
+    }
+}