+++ Assets/Dust/Experimental/Scripts/Runtime/Variables/Managers/ObjectVariablesManager.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Dust.Experimental.Variables
+{
+    public class ObjectVariablesManager : AbstractVariablesManager
+    {
+    }
+}